Update network activity indicator handling
Added some concurrency support. If you were to decrement the request
count down to zero, then increment it within 0.5 seconds, the activity
indicator would turn off 0.5 seconds after the decrement.

This implementation uses a timer to delay the hiding (which we want to
avoid flickering), but when the function gets called, it will recheck
the request count, and then act depending on what it is at that time.